--- packages/color/jest.config.cjs
@@ -1,14 +1,0 @@
-module.exports = {
-  clearMocks: true,
-  collectCoverage: true,
-  collectCoverageFrom: ['src/**/*.js'],
-  coverageDirectory: 'coverage',
-  coveragePathIgnorePatterns: ['<rootDir>/dist/', '<rootDir>/test/'],
-  coverageReporters: [['lcov', { projectRoot: '../..' }], 'text', 'clover'],
-  errorOnDeprecated: true,
-  testEnvironment: 'node',
-  testPathIgnorePatterns: ['<rootDir>/dist/'],
-  transform: {
-    '^.+\\.(t|j)sx?$': ['@swc/jest', { configFile: '../../.swcrc.test' }],
-  },
-};
