--- packages/color/package.json
@@ -1,52 +1,0 @@
-{
-  "name": "@lowdefy/color",
-  "version": "3.23.1",
-  "license": "Apache-2.0",
-  "description": "",
-  "homepage": "https://lowdefy.com",
-  "keywords": [
-    "lowdefy",
-    "utils"
-  ],
-  "bugs": {
-    "url": "https://github.com/lowdefy/lowdefy/issues"
-  },
-  "contributors": [
-    {
-      "name": "Sam Tolmay",
-      "url": "https://github.com/SamTolmay"
-    },
-    {
-      "name": "Gerrie van Wyk",
-      "url": "https://github.com/Gervwyk"
-    }
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/lowdefy/lowdefy.git"
-  },
-  "type": "module",
-  "exports": "./dist/index.js",
-  "files": [
-    "dist/*"
-  ],
-  "scripts": {
-    "build": "yarn swc",
-    "clean": "rm -rf dist",
-    "prepare": "yarn build",
-    "swc": "swc src --out-dir dist --config-file ../../.swcrc --delete-dir-on-start",
-    "test": "jest --coverage"
-  },
-  "dependencies": {
-    "tinycolor2": "1.4.2"
-  },
-  "devDependencies": {
-    "@swc/cli": "0.1.51",
-    "@swc/core": "1.2.107",
-    "@swc/jest": "0.2.5",
-    "jest": "27.3.1"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
