+++ packages/plugins/blocks/blocks-antd/package.json
@@ -47,7 +47,6 @@
   "dependencies": {
     "@ant-design/icons": "4.7.0",
     "@lowdefy/block-utils": "3.22.0",
-    "@lowdefy/color": "3.22.0",
     "@lowdefy/helpers": "3.22.0",
     "antd": "4.17.0-alpha.7",
     "classnames": "2.3.1",
@@ -55,7 +54,7 @@
     "rc-animate": "3.1.1",
     "react": "18.0.0-alpha-327d5c484-20211106",
     "react-dom": "18.0.0-alpha-327d5c484-20211106",
-    "react-router-dom": "5.2.1"
+    "tinycolor2": "1.4.2"
   },
   "devDependencies": {
     "@babel/core": "7.16.0",
