+++ packages/plugins/blocks/blocks-antd/src/blocks/Button/Button.js
@@ -16,9 +16,10 @@
 
 import React from 'react';
 import { Button } from 'antd';
-import color from '@lowdefy/color';
 import { get, type } from '@lowdefy/helpers';
 import { blockDefaultProps, renderHtml } from '@lowdefy/block-utils';
+
+import color from '../../color.js';
 
 const ButtonBlock = ({
   blockId,
